# docs/api_spec.yaml
openapi: 3.0.3
info:
  title: 商用地理信息虚假内容检查与反投喂平台 API
  description: |
    ## 概述
    本API提供地理信息虚假内容检测服务，支持单条/批量检测、规则管理、审核标注、统计报表等功能。
    
    ## 认证方式
    - **JWT Token**: 用于Web端用户认证，在Header中添加 `Authorization: Bearer <token>`
    - **API Key**: 用于服务端对接，在Header中添加 `X-API-Key: <api_key>`
    
    ## 请求限制
    - 基础版: 10,000次/天
    - 专业版: 50,000次/天
    - 企业版: 不限制
    
    ## 错误码
    | 错误码 | 说明 |
    |--------|------|
    | 400 | 请求参数错误 |
    | 401 | 未授权/Token过期 |
    | 403 | 权限不足 |
    | 404 | 资源不存在 |
    | 429 | 请求频率超限 |
    | 500 | 服务器内部错误 |
    
  version: 1.0.0
  contact:
    name: API Support
    email: api@geo-fake-detection.com
servers:
  - url: https://api.geo-fake-detection.com/v1
    description: 生产环境
  - url: https://api-staging.geo-fake-detection.com/v1
    description: 测试环境
...
paths:
...
  /records:
    get:
      tags: [Records]
      summary: 记录列表
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: device_id
          in: query
          schema:
            type: string
        - name: user_id
          in: query
          schema:
            type: string
        - name: risk_level
          in: query
          schema:
            type: string
            enum: [low, medium, high]
        - name: review_status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, escalated]
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
        - name: keyword
          in: query
          description: 内容关键词搜索
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
  /records/{id}:
    get:
      tags: [Records]
      summary: 记录详情
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetectionRecord'

    delete:
      tags: [Records]
      summary: 删除记录
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /records/{id}/label:
    put:
      tags: [Records]
      summary: 标注记录
      description: 为记录添加真实标签，用于模型训练
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ground_truth]
              properties:
                ground_truth:
                  type: boolean
                  description: 真实标签
                feedback_type:
                  type: string
                  enum: [false_positive, false_negative, correct]
                feedback_comment:
                  type: string
      responses:
        '200':
          description: 标注成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /records/export:
    post:
      tags: [Records]
      summary: 导出记录
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                format:
                  type: string
                  enum: [csv, json, excel]
                  default: csv
                filters:
                  type: object
                fields:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: 导出成功
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary

  # ============================================
  # 审核接口
  # ============================================
...
  /keywords:
    get:
      tags: [Keywords]
      summary: 关键词列表
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: enabled
          in: query
          schema:
            type: boolean
        - name: keyword
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Keyword'

    post:
      tags: [Keywords]
      summary: 添加关键词
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KeywordCreate'
      responses:
        '200':
          description: 添加成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Keyword'
  /keywords/{id}:
    put:
      tags: [Keywords]
      summary: 更新关键词
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KeywordUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Keyword'

    delete:
      tags: [Keywords]
      summary: 删除关键词
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /keywords/import:
    post:
      tags: [Keywords]
      summary: 批量导入关键词
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                category:
                  type: string
      responses:
        '200':
          description: 导入成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported_count:
                    type: integer
                  skipped_count:
                    type: integer
  /keywords/export:
    get:
      tags: [Keywords]
      summary: 导出关键词
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 导出成功
          content:
            text/csv:
              schema:
                type: string

  # ============================================
  # 统计报表
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
    SuccessResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
...
    PaginatedResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: array
          items:
            type: object
        pagination:
          type: object
          properties:
            page:
              type: integer
            page_size:
              type: integer
            total:
              type: integer
            total_pages:
              type: integer
...
    DetectionRecord:
      type: object
      properties:
        id:
          type: integer
        record_id:
          type: string
        device_id:
          type: string
        user_id:
          type: string
        timestamp:
          type: integer
        latitude:
          type: number
        longitude:
          type: number
        content_text:
          type: string
        content_type:
          type: string
        suspicion_score:
          type: number
        risk_level:
          type: string
        is_fake:
          type: boolean
        geo_score:
          type: number
        text_score:
          type: number
        simhash_score:
          type: number
        semantic_score:
          type: number
        reasons:
          type: array
          items:
            type: string
        review_status:
          type: string
        ground_truth:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    Keyword:
      type: object
      properties:
        id:
          type: integer
        category:
          type: string
        keyword:
          type: string
        weight:
          type: number
        risk_level:
          type: string
        source:
          type: string
        hit_count:
          type: integer
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
    KeywordCreate:
      type: object
      required: [category, keyword]
      properties:
        category:
          type: string
        keyword:
          type: string
        weight:
          type: number
          default: 1.0
        risk_level:
          type: string
          default: medium
        description:
          type: string
    KeywordUpdate:
      type: object
      properties:
        weight:
          type: number
        risk_level:
          type: string
        enabled:
          type: boolean
...